Changed constructor parameter to Int16

+++ BookDescription/Year.swift
@@ -9,7 +9,7 @@ struct Year : CustomStringConvertible {
     
     var description: String = "SORRY, NOT IMPLEMENTED YET"
     
-    init(yyyy: Int) {
+    init(yyyy: Int16) {
         //
     }
